// src/__generated__/deleted-documents-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type DeletedDocumentsQuery = {
    documents: ({
        __typename: "ES4Document" | "UnknownFormatDocument";
    })[];
    workspaces: ({
        documents: ({
            __typename: "ES4Document" | "UnknownFormatDocument";
        })[];
    })[];
    authors: ({
        documents: ({
            __typename: "ES4Document" | "UnknownFormatDocument";
        })[];
    })[];
};
export type DeletedDocumentsQueryVariables = {
    includeDeleted: boolean;
};