// src/__generated__/long-name-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type LongNameQuery = {
    authors: ({
        displayName: string | null;
    })[];
    workspaces: ({
        address: string;
        authors: ({
            displayName: string | null;
        })[];
    })[];
};
export type LongNameQueryVariables = {};